backend/requirements.txt
pydantic
litellm
flask
flask-cors
python-dotenv
supermemory